feat: update services management

+++ src/app/enterprise/services/pages/services-management/services-management.component.ts
@@ -9,12 +9,15 @@ import { EntsevicesService } from "../../services/entsevices.service";
 import { Entservice } from "../../model/entservice.entity";
 import { ServicesCreateAndEditComponent } from "../../components/services-create-and-edit/services-create-and-edit.component";
 import { NgClass } from "@angular/common";
-import { TranslateModule } from "@ngx-translate/core";
+import {TranslateModule, TranslateService} from "@ngx-translate/core";
+import {MatMenuModule} from "@angular/material/menu";
+import {MatButtonModule} from "@angular/material/button";
 
 @Component({
   selector: 'app-services-management',
   standalone: true,
-  imports: [MatPaginator, MatSort, MatIconModule, ServicesCreateAndEditComponent, MatTableModule, NgClass, TranslateModule],
+  imports: [MatPaginator, MatSort, MatIconModule, ServicesCreateAndEditComponent, MatTableModule,
+    NgClass, TranslateModule, MatMenuModule, MatButtonModule],
   templateUrl: './services-management.component.html',
   styleUrl: './services-management.component.css'
 })
@@ -23,15 +26,17 @@ export class ServicesManagementComponent implements OnInit, AfterViewInit{
   // Attributes
   entservicesData: Entservice;
   dataSource!: MatTableDataSource<any>;
-  displayedColumns: string[] = ['id', 'name', 'categoryId', 'imageUrl', 'description', 'basePrice', 'beautySalonId', 'actions'];
+  displayedColumns: string[] = ['table-number', 'name', 'categoryId', 'imageUrl', 'description', 'basePrice', 'actions'];
   isEditMode: boolean;
   totalServices: number;
 
   @ViewChild(MatPaginator, { static: false}) paginator!: MatPaginator;
   @ViewChild(MatSort, { static: false}) sort!: MatSort;
 
+  @ViewChild(ServicesCreateAndEditComponent) createEditComponent!: ServicesCreateAndEditComponent;
+
   // Constructor
-  constructor(private entserviceService: EntsevicesService) {
+  constructor(private entserviceService: EntsevicesService, private translate: TranslateService) {
     this.isEditMode = false;
     this.entservicesData = {} as Entservice;
     this.dataSource = new MatTableDataSource<any>();
@@ -46,17 +51,27 @@ export class ServicesManagementComponent implements OnInit, AfterViewInit{
 
   // CRUD Actions
 
+  /*
   private getAllEntservices(): void {
     this.entserviceService.getAll()
       .subscribe((response: any) => {
         this.dataSource.data = response;
         console.log("data for table", this.dataSource.data);
-
       });
   };
+  */
+
+  //Get services by salon id
+  private getServicesBySalon(): void {
+    this.entserviceService.getAllServicesBySalonId()
+      .subscribe((response: any) => {
+        this.dataSource.data = response;
+        console.log("data for salon", this.dataSource.data);
+      })
+  }
 
   private getTotalServices(): any {
-    this.entserviceService.getAll()
+    this.entserviceService.getAllServicesBySalonId()
       .subscribe((response: any) => {
         this.totalServices = response.length;
         console.log('Total services:', this.totalServices)
@@ -64,11 +79,15 @@ export class ServicesManagementComponent implements OnInit, AfterViewInit{
       });
   };
 
+  // Additional functions
+
   private createEntservice(): void {
+    // Set salon id
+    this.entservicesData.beauty_salon_id = this.entserviceService.getSalonContextId();
     this.entserviceService.create(this.entservicesData)
       .subscribe((response: any) => {
         this.dataSource.data.push({...response});
-        // Actualiza el dataSource.data con los students actuales, para que Angular detecte el cambio y actualice la vista.
+        // Actualiza el dataSource.data con los service actuales, para que Angular detecte el cambio y actualice la vista.
         this.dataSource.data = this.dataSource.data
           .map((service: Entservice) => {
             return service;
@@ -103,23 +122,41 @@ export class ServicesManagementComponent implements OnInit, AfterViewInit{
   // UI Event Handlers
 
   onEditItem(element: Entservice) {
+    this.translate.get('bc-4.addEdit').subscribe((translatedMessage: string) =>
+    { window.alert(translatedMessage); })
     this.isEditMode = true;
     this.entservicesData = element;
   }
 
   onDeleteItem(element: Entservice) {
-    this.deleteEntservice(element.id);
+    this.getTotalServices();
+    if (this.totalServices === 1){
+      this.translate.get('bc-4.addIfOne').subscribe((translatedMessage: string) =>
+      window.alert(translatedMessage));
+    } else {
+      this.translate.get('bc-4.addCancel').subscribe((translatedMessage: string) =>
+      {const result = window.confirm(translatedMessage);
+        if (result){
+          this.deleteEntservice(element.id);
+          this.getTotalServices();
+          this.createEditComponent.loadExistingNames();
+        }
+      })
+    }
   }
 
   onCancelEdit() {
     this.resetEditState();
-    this.getAllEntservices();
+    this.getServicesBySalon();
   }
 
   onEntserviceAdded(element: Entservice) {
     this.entservicesData = element;
     this.createEntservice();
     this.resetEditState();
+    // Update amount after create a service
+    this.getTotalServices();
+    this.createEditComponent.loadExistingNames();
   }
 
   onEntserviceUpdated(element: Entservice) {
@@ -127,10 +164,6 @@ export class ServicesManagementComponent implements OnInit, AfterViewInit{
     this.updateEntservice();
     this.resetEditState();
   }
-
-  // Additional functions
-
-
 
   // Lifecycle Hooks
 
@@ -140,12 +173,10 @@ export class ServicesManagementComponent implements OnInit, AfterViewInit{
   }
 
   ngOnInit(): void {
-    this.getAllEntservices();
+    // Get total amount of services
     this.getTotalServices();
+    this.getServicesBySalon()
   }
-
-
-
 
 
 }
